Split measurements by type and added an abstract superclass. Point to point measurement has it's own implementation now
